hotfix : 공동구매 작성자의 참여 취소 시 예외 처리
- 클라가 서버 붙이다가 작성자 본인이 요청한 경우가 있었어서 예외처리했어유
- 참여 취소해도 덧글은 안 날리길래, commentRepository.deleteById도 한 줄 넣었어유

+++ src/main/java/com/zeepy/server/community/service/CommunityService.java
@@ -11,6 +11,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.zeepy.server.common.CustomExceptionHandler.CustomException.AlreadyParticipationException;
 import com.zeepy.server.common.CustomExceptionHandler.CustomException.BadRequestCommentException;
+import com.zeepy.server.common.CustomExceptionHandler.CustomException.JointPurchaseOwner;
 import com.zeepy.server.common.CustomExceptionHandler.CustomException.MoreThanOneParticipantException;
 import com.zeepy.server.common.CustomExceptionHandler.CustomException.NotFoundCommunityException;
 import com.zeepy.server.common.CustomExceptionHandler.CustomException.NotFoundUserException;
@@ -168,6 +169,10 @@ public class CommunityService {
 		Community findCommunity = communityRepository.findById(communityId)
 			.orElseThrow(NotFoundCommunityException::new);
 
+		if (findUser.equals(findCommunity.getUser())) {
+			throw new JointPurchaseOwner();
+		}
+
 		Long findUserId = findUser.getId();
 		Long findCommunityId = findCommunity.getId();
 		participationRepository.deleteByUserIdAndCommunityId(findUserId, findCommunityId);
@@ -179,8 +184,8 @@ public class CommunityService {
 			.findFirst()
 			.orElseThrow(BadRequestCommentException::new);
 		superComment.cancelParticipation();
-
 		findCommunity.substractCurrentNumberOfPeople();
+		commentRepository.deleteById(superComment.getId());
 
 		/**
 		 * PUSH 알림 추가 로직
